+++ src/data_preprocessing/fc_preprocessing.py
@@ -13,12 +13,12 @@ def get_train_test(df, window=12, train_size=0.8):
     for j in range(len(df) - window):
         X.append(df[j: j + window])
         y.append(df["price actual"].iloc[j + window])
-    X = np.array(X)
+    X = np.array(X).astype(np.float32)
     X = X.reshape(len(X), -1)
-    y = np.array(y)
+    y = np.array(y).astype(np.float32)
     x_train = X[:int(len(X) * train_size)]
     y_train = y[:int(len(y) * train_size)]
-    valid_size = int(len(X) - len(x_train) / 2)
+    valid_size = int((len(X) - len(x_train)) / 2)
     x_valid = X[int(len(X) * train_size): int(len(X) * train_size) + valid_size]
     y_valid = y[int(len(y) * train_size): int(len(y) * train_size) + valid_size]
     x_test = X[int(len(X) * train_size) + valid_size:]
